+++ exchange-wrapper/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@rarible/exchange-wrapper",
+  "version": "0.1.0",
+  "main": "truffle.js",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "test": "truffle test"
+  },
+  "dependencies": {
+    "@openzeppelin/contracts-upgradeable": "3.4.0",
+    "@rarible/exchange-interfaces": "0.2.0",
+    "@rarible/lib-asset": "^0.1.2",
+    "@rarible/libraries": "^0.1.2",
+    "@rarible/royalties": "^0.7.4",
+    "@rarible/royalties-upgradeable": "^0.7.4",
+    "eth-sig-util": "^3.0.1"
+  },
+  "devDependencies": {
+    "@daonomic/tests-common": "0.2.2",
+    "@openzeppelin/truffle-upgrades": "^1.8.0",
+    "@truffle/hdwallet-provider": "1.4.0",
+    "sol-merger": "^3.1.0",
+    "truffle-assertions": "0.9.2",
+    "truffle-plugin-verify": "0.5.7"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "gitHead": "c149f3e740650a91e0764245230dbddc8f8f423e"
+}
