+++ exchange-wrapper/test/contracts/v2/ExchangeV2Test.sol
@@ -1,0 +1,24 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity 0.7.6;
+pragma abicoder v2;
+
+import "@rarible/exchange-v2/contracts/ExchangeV2Core.sol";
+import "@rarible/exchange-v2/contracts/RaribleTransferManager.sol";
+import "@rarible/royalties/contracts/IRoyaltiesProvider.sol";
+
+contract ExchangeV2Test is ExchangeV2Core, RaribleTransferManager {
+    function __ExchangeV2_init(
+        INftTransferProxy _transferProxy,
+        IERC20TransferProxy _erc20TransferProxy,
+        uint newProtocolFee,
+        address newDefaultFeeReceiver,
+        IRoyaltiesProvider newRoyaltiesProvider
+    ) external initializer {
+        __Context_init_unchained();
+        __Ownable_init_unchained();
+        __TransferExecutor_init_unchained(_transferProxy, _erc20TransferProxy);
+        __RaribleTransferManager_init_unchained(newProtocolFee, newDefaultFeeReceiver, newRoyaltiesProvider);
+        __OrderValidator_init_unchained();
+    }
+}
