+++ exchange-wrapper/test/order.js
@@ -1,0 +1,76 @@
+const EIP712 = require("./EIP712");
+
+function AssetType(assetClass, data) {
+	return { assetClass, data }
+}
+
+function Asset(assetClass, assetData, value) {
+	return { assetType: AssetType(assetClass, assetData), value };
+}
+
+function Order(maker, makeAsset, taker, takeAsset, salt, start, end, dataType, data) {
+	return { maker, makeAsset, taker, takeAsset, salt, start, end, dataType, data };
+}
+
+function OpenSeaOrdersInput(
+  addrs,
+  uints,
+  feeMethodsSidesKindsHowToCalls,
+  calldataBuy,
+  calldataSell,
+  replacementPatternBuy,
+  replacementPatternSell,
+  staticExtradataBuy,
+  staticExtradataSell,
+  vs,
+  rssMetadata) {
+	return {
+	  addrs,
+	  uints,
+	  feeMethodsSidesKindsHowToCalls,
+	  calldataBuy,
+	  calldataSell,
+	  replacementPatternBuy,
+	  replacementPatternSell,
+	  staticExtradataBuy,
+	  staticExtradataSell,
+	  vs,
+	  rssMetadata };
+}
+
+function TradeData(marketWyvern, amount, tradeData) {return {marketWyvern, amount, tradeData};};
+
+const Types = {
+	AssetType: [
+		{name: 'assetClass', type: 'bytes4'},
+		{name: 'data', type: 'bytes'}
+	],
+	Asset: [
+		{name: 'assetType', type: 'AssetType'},
+		{name: 'value', type: 'uint256'}
+	],
+	Order: [
+		{name: 'maker', type: 'address'},
+		{name: 'makeAsset', type: 'Asset'},
+		{name: 'taker', type: 'address'},
+		{name: 'takeAsset', type: 'Asset'},
+		{name: 'salt', type: 'uint256'},
+		{name: 'start', type: 'uint256'},
+		{name: 'end', type: 'uint256'},
+		{name: 'dataType', type: 'bytes4'},
+		{name: 'data', type: 'bytes'},
+	]
+};
+
+async function sign(order, account, verifyingContract) {
+	const chainId = Number(await web3.eth.getChainId());
+	const data = EIP712.createTypeData({
+		name: "Exchange",
+		version: "2",
+		chainId,
+		verifyingContract
+	}, 'Order', order, Types);
+	return (await EIP712.signTypedData(web3, account, data)).sig;
+}
+
+module.exports = { AssetType, Asset, Order, OpenSeaOrdersInput, TradeData, sign }
