+++ exchange-wrapper/truffle-config.js
@@ -1,0 +1,1 @@
+module.exports = require('../truffle-config.js');
