+++ package.json
@@ -9,6 +9,7 @@
             "custom-matchers",
             "deploy",
             "exchange-v2",
+            "exchange-wrapper",
             "interfaces",
             "meta-tx",
             "lazy-mint",
